+++ CHANGELOG.md
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 ### Fixed
-- Although the custom garrison query was working, it was slightly configured incorrectly in how activities were retrieved. The query now heavily filters based off the current node which should speed up querying.
+- Although the custom garrison query was working, it was slightly configured incorrectly in how it retrieved activities. The query now heavily filters based off the current node.
 
 ## [0.19.8] - 2026-03-02