+++ CHANGELOG.md
@@ -5,10 +5,12 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 ### Added
+- Pool status for the health check endpoint
 
 ### Changed
 
 ### Fixed
+- Although the custom garrison query was working, it was slightly configured incorrectly in how it retrieved activities. The query now heavily filters based off the current node.
 
 ## [0.19.8] - 2026-03-02