+++ CHANGELOG.md
@@ -5,9 +5,15 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 ### Added
-- Pool status for the health check endpoint
 
 ### Changed
+
+### Fixed
+
+## [0.19.9] - 2026-03-09
+
+### Added
+- Pool status for the health check endpoint
 
 ### Fixed
 - Although the custom garrison query was working, it was slightly configured incorrectly in how it retrieved activities. The query now heavily filters based off the current node.